# BUG: transform and filter misbehave when grouping on categorical data