--- a/doc/source/whatsnew/v0.16.1.txt
+++ b/doc/source/whatsnew/v0.16.1.txt
@@ -247,3 +247,5 @@ Bug Fixes
 
 
 - Bug in hiding ticklabels with subplots and shared axes when adding a new plot to an existing grid of axes (:issue:`9158`)
+- Bug in ``transform`` and ``filter`` when grouping on a categorical variable (:issue:`9921`)
+- Bug in ``transform`` when groups are equal in number and dtype to the input index (:issue:`9700`)