Fixes GH #9921
closes #9700