@evanpw 2015-04-28
Conditional is cleaned up, squashed, and tests are green.